Write test for delete doctor that throw DoctorNotFoundException when doctor with given id that not exist

+++ src/DDTDoctorAppointment.Services/Doctors/DoctorAppService.cs
@@ -70,6 +70,10 @@ namespace DDTDoctorAppointment.Services.Doctors
         public void Delete(int id)
         {
             var doctor = _repository.Getbyid(id);
+            if (doctor==null)
+            {
+                throw new DoctorNotFoundException();
+            }
 
             _repository.Remove(doctor);
